bothub-nlp-nlu-worker/requirements.txt
-i https://pypi.org/simple
plac==0.9.6
psycopg2-binary==2.7.6.1
requests==2.20.1
contextvars==2.3
spacy==2.0.18
rasa-nlu==0.14.1
tensorflow==1.12.0
scikit-learn==0.20.2
sklearn-crfsuite==0.3.6
pythainlp==1.7.2
pymorphy2==0.8
plac==0.9.6
requests==2.20.1
python-decouple==3.1
celery==4.2.1
git+https://github.com/dyohan9/get_backend_python.git@master